Move a version-specific extractors to trait
